+++ c++/pod5_format/run_info_table_writer.h
@@ -82,7 +82,7 @@ private:
 /// \returns The writer for the new table.
 POD5_FORMAT_EXPORT Result<RunInfoTableWriter> make_run_info_table_writer(
     std::shared_ptr<arrow::io::OutputStream> const & sink,
-    std::shared_ptr<const arrow::KeyValueMetadata> const & metadata,
+    std::shared_ptr<arrow::KeyValueMetadata const> const & metadata,
     std::size_t table_batch_size,
     arrow::MemoryPool * pool);